apps/rag-server/pyproject.toml
[project]
name = "rag-server"
version = "0.1.0"
description = "FastAPI + ChromaDB + Google Gemini RAG API Server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "google-generativeai>=0.5.4",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
